Fix inaccuracy in documentation comments (sorry for the tiny commit)

+++ src/movie.cpp
@@ -1369,10 +1369,10 @@ bool FCEUMOV_ReadState(EMUFILE* is, uint32 size)
 
 			if (sameTimeline)
 			{
-				//if we made it this far, then the savestate has identical movie data but we want to know now if the stateMOVIE size is greater than current movie size and make this error
+				//if we made it this far, then the savestate has identical movie data but we want to know now if the state frame count is greater than current movie size and make this error
 				
 				//currFrameCounter at this point represents the savestate framecount
-				if(currFrameCounter > currMovieData.records.size()) 
+				if(currFrameCounter > (int)currMovieData.records.size()) 
 				{
 					//TODO: turn frame counter to red to get attention
